is2: Importance Sampling Squared

The is2 package provides a general purpose implementation of the marginal likelihood calculation methods outlined in Tran et al. (2020) doi: 10.3758/s13428-020-01348-w . This package will work with pmwg sampler objects, and can be also be used with the output of other MCMC samplers.

Documentation

The documentation found at https://newcastlecl.github.io/is2/ and contains background information and motivation for the approach used in this package and several detailed examples of the package in action.

User input

The user is expected to provide a structured object with MCMC samples, a log likelihood function. The initial version works well with a pmwg sampler object.

References

Tran, M. N., Scharth, M., Gunawan, D., Kohn, R., Brown, S. D., & Hawkins, G. E. (2020). Robustly estimating the marginal likelihood for cognitive models via importance sampling. Behavior Research Methods, 1-18.

See also

Author

Maintainer: Gavin Cooper gavin@gavincooper.net (Package creator and maintainer) [translator]

Authors:

Reilly Innes Reilly.Innes@uon.edu.au

Caroline Kuhne caroline.kuhne@newcastle.edu.au

Jon-Paul Cavallaro jon-paul.cavallaro@uon.edu.au

David Gunawan dgunawan@uow.edu.au (Author of original MATLAB code)

Guy Hawkins guy.hawkins@newcastle.edu.au

Scott Brown scott.brown@newcastle.edu.au (Original translation from MATLAB to R) [translator]